reorganize sage-api crate
